==> 01-TratamentoInicial.ipynb <==

#Sobrevivência empresarial em Pernambuco com base no PIB e nos dados abertos da RFB: uma análise em dois períodos sociopolítico econômicos da história do Brasil


#Elisângela P.M. Santana¹, Edivaldo R.S. Júnior², Cleide Lins de Almeida³


##SCRIPT01 - TRATAMENTO DE DADOS INICIAL

BAIXAR A BASE DE DADOS DA RECEITA (Dados Públicos CNPJ) \
https://www.gov.br/receitafederal/pt-br/assuntos/orientacao-tributaria/cadastros/consultas/dados-publicos-cnpj. Acesso em 01/05/2022

BASES UTILIZADAS: \
Dados Abertos CNPJ EMPRESA 01 > K3241.K03200Y0.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 02 > K3241.K03200Y1.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 03 > K3241.K03200Y2.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 04 > K3241.K03200Y3.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 05 > K3241.K03200Y4.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 06 > K3241.K03200Y5.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 07 > K3241.K03200Y6.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 08 > K3241.K03200Y7.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 09 > K3241.K03200Y8.D20312.EMPRECSV \
Dados Abertos CNPJ EMPRESA 10 > K3241.K03200Y9.D20312.EMPRECSV \
Dados Abertos CNPJ ESTABELECIMENTO 01 > K3241.K03200Y0.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 02 > K3241.K03200Y1.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 03 > K3241.K03200Y2.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 04 > K3241.K03200Y3.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 05 > K3241.K03200Y4.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 06 > K3241.K03200Y5.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 07 > K3241.K03200Y6.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 08 > K3241.K03200Y7.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 09 > K3241.K03200Y8.D20312.ESTABELE \
Dados Abertos CNPJ ESTABELECIMENTO 10 > K3241.K03200Y9.D20312.ESTABELE \
Informações sobre o Simples Nacional/MEI > F.K03200$W.SIMPLES.CSV.D20312 \
Tabela de atributo CNAE > F.K03200$Z.D20312.CNAECSV \
Tabela de motivo da situação cadastral > F.K03200$Z.D20312.MOTICSV \
Tabela de atributo Município > F.K03200$Z.D20312.MUNICCSV \
Tabela de atributo Natureza Jurídica > F.K03200$Z.D20312.NATJUCSV \
Tabela de atributo País > F.K03200$Z.D20312.PAISCSV \
Tabela de atributo qualificação dos sócios > F.K03200$Z.D20312.QUALSCSV \
Dados Abertos Sítio RFB Data Base 13-03-202222 > Extracao 01-05-2022


# CARREGANDO AS BIBLIOTECAS 

#!pip install matplotlib 
#!pip install numpy
#!pip install pandas
%matplotlib inline
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd


#VISUALIZAR PACOTES INSTALADOS
#!pip freeze

# CRIAR ARQUIVO REQUERIMENTS.TXT
#!pip freeze > requirements.txt

caminho = 'D:\\#PROJETOS\\FURG\\ECD-CIENCIA DE DADOS\\#TCC-FURG\\BASE-RFB-12-03-2022'
tratada = 'D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA'

IMPORTAR TABELAS COMPLEMENTARES (SIMPLES , CNAE, MOTIVO, MUNICIPIO, NATUREZA JURIDICA, PAIS, QUALIFICACAO)

#IMPORTAR TABELA SIMPLES NACIONAL (simples) #DIRETORIOS DEFINIDOS
os.chdir(caminho)


simples = pd.read_csv('F.K03200$W.SIMPLES.CSV.D20312.csv', sep=';',header=None)
simples.columns = ['CNPJ BASICO', 'OPÇÃO PELO SIMPLES','DATA_INC_SIM','DATA EXC_SIMP','MEI','DATA_INT_MEI','DATA_EXC_MEI']
simples.head()

#EXPORTAR TABELA SIMPLES NACIONAL
os.chdir(tratada)
simples.to_csv(r'simples.csv', index = False)



simples.info()

#LIMPAR MEMORIA
del simples
import gc
gc.collect()

#IMPORTAR TABELA CNAE (cnae)
os.chdir(caminho)
cnae = pd.read_csv('F.K03200$Z.D20312.CNAECSV.csv', sep=";" , encoding='ISO-8859-1',header=None)
cnae.columns = ['N_CNAE', 'DESC_CNAE']
cnae.head()

#EXPORTAR TABELA CNAE
os.chdir(tratada)
cnae.to_csv(r'cnae.csv', index = False)

cnae.info()

#LIMPAR MEMORIA
del cnae
import gc
gc.collect()

#IMPORTAR TABELA MOTIVO (motivo)
os.chdir(caminho)
motivo = pd.read_csv('F.K03200$Z.D20312.MOTICSV.csv', sep=";" , encoding='ISO-8859-1', header=None)
motivo.columns = ['N_MOTIVO', 'DESC_MOTIVO']
motivo.head()

#EXPORTAR TABELA MOTIVO
os.chdir(tratada)
motivo.to_csv(r'motivo.csv', index = False)

motivo.info()

#LIMPAR MEMORIA
del motivo
import gc
gc.collect()

#IMPORTAR TABELA MUNICIPIO (municipio)
os.chdir(caminho)
municipio = pd.read_csv('F.K03200$Z.D20312.MOTICSV.csv', sep=";" , encoding='ISO-8859-1',header=None)
municipio.columns = ['N_MUNIC', 'DESC_MUNIC']
municipio.head()

#EXPORTAR TABELA MUNICIPIO (municipio)
os.chdir(tratada)
municipio.to_csv(r'municipio.csv', index = False)

municipio.info()

#LIMPAR MEMORIA
del municipio
import gc
gc.collect()

#IMPORTAR TABELA NATUREZA JURIDICA (natju)
os.chdir(caminho)
natju = pd.read_csv('F.K03200$Z.D20312.NATJUCSV.csv', sep=";" , encoding='ISO-8859-1',header=None)
natju.columns = ['N_NATJU', 'DESC_NATJU']
natju.head()

#EXPORTAR TABELA NATJU
os.chdir(tratada)
natju.to_csv(r'natju.csv', index = False)

natju.info()

#LIMPAR MEMORIA
del natju
import gc
gc.collect()

#IMPORTAR TABELA PAIS (pais)
os.chdir(caminho)
pais = pd.read_csv('F.K03200$Z.D20312.PAISCSV.csv', sep=";" , encoding='ISO-8859-1')
pais.columns = ['N_PAIS', 'DESC_PAIS']
pais.head()

#EXPORTAR TABELA PAIS
os.chdir(tratada)
pais.to_csv(r'pais.csv', index = False)

pais.info()

#LIMPAR MEMORIA
del pais
import gc
gc.collect()

#IMPORTAR TABELA QUALIFICACAO (qualif)
os.chdir(caminho)
qualif = pd.read_csv('F.K03200$Z.D20312.QUALSCSV.csv', sep=';', encoding='ISO-8859-1')
qualif.columns = ['N_QUALIF', 'DESC_QUALIF']
qualif.head()

#EXPORTAR TABELA QUALIFICAÇÃO
os.chdir(tratada)
qualif.to_csv(r'qualif.csv', index = False)

qualif.info()

#LIMPAR MEMORIA
del qualif
import gc
gc.collect()

TRATAMENTO DE DADOS INICIAL ARQUIVOS EMPRESA SITE DA RFB  K3241.K03200Y0.D20312.EMPRE.CSV  A K3241.K03200Y9.D20312.EMPRE.CSV

#IMPORTAR TABELA EMPRESAS #00
os.chdir(caminho)
empresa0 = pd.read_csv('K3241.K03200Y0.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa0.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']
empresa0.head()

#EXPORTAR TABELA EMPRESA_0
os.chdir(tratada)
empresa0.to_csv(r'empresa0.csv', index = False)

#LIMPAR MEMORIA
del empresa0
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #01 
os.chdir(caminho)
empresa1 = pd.read_csv('K3241.K03200Y1.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa1.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_1
os.chdir(tratada)
empresa1.to_csv(r'empresa1.csv', index = False)

#LIMPAR MEMORIA
del empresa1
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #02 
os.chdir(caminho)
empresa2 = pd.read_csv('K3241.K03200Y2.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa2.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_2
os.chdir(tratada)
empresa2.to_csv(r'empresa2.csv', index = False)

#LIMPAR MEMORIA
del empresa2
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #03 
os.chdir(caminho)
empresa3 = pd.read_csv('K3241.K03200Y3.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa3.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_3
os.chdir(tratada)
empresa3.to_csv(r'empresa3.csv', index = False)

#LIMPAR MEMORIA
del empresa3
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #04 
os.chdir(caminho)
empresa4 = pd.read_csv('K3241.K03200Y4.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa4.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_4
os.chdir(tratada)
empresa4.to_csv(r'empresa4.csv', index = False)

#LIMPAR MEMORIA
del empresa4
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #05 
os.chdir(caminho)
empresa5 = pd.read_csv('K3241.K03200Y5.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa5.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_5
os.chdir(tratada)
empresa5.to_csv(r'empresa5.csv', index = False)

#LIMPAR MEMORIA
del empresa5
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #06 
os.chdir(caminho)
empresa5 = pd.read_csv('K3241.K03200Y6.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa5.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_6
os.chdir(tratada)
empresa5.to_csv(r'empresa6.csv', index = False)

#LIMPAR MEMORIA
del empresa5
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #07 
os.chdir(caminho)
empresa5 = pd.read_csv('K3241.K03200Y7.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa5.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_7
os.chdir(tratada)
empresa5.to_csv(r'empresa7.csv', index = False)

#LIMPAR MEMORIA
del empresa5
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #08 
os.chdir(caminho)
empresa8 = pd.read_csv('K3241.K03200Y8.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa8.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_8
os.chdir(tratada)
empresa8.to_csv(r'empresa8.csv', index = False)

#LIMPAR MEMORIA
del empresa8
import gc
gc.collect()

#IMPORTAR TABELA EMPRESAS #09 
os.chdir(caminho)
empresa9 = pd.read_csv('K3241.K03200Y9.D20312.EMPRECSV.CSV', sep=';', encoding='ISO-8859-1')
empresa9.columns = ['CNPJ BASICO', 'RAZÃO SOCIAL','NATJU','QUALIF','CAP_SOCIAL','PORTE','ENTE_FED']

#EXPORTAR TABELA EMPRESA_9
os.chdir(tratada)
empresa9.to_csv(r'empresa9.csv', index = False)

#LIMPAR MEMORIA
del empresa9
import gc
gc.collect()

os.chdir(tratada)

#IMPORTAR DATAFRAME EMPRESA (COLUNAS SELECIONADAS)
empresa0 = pd.read_csv('empresa0.csv', sep=',', encoding='UTF-8')
empresa1 = pd.read_csv('empresa1.csv', sep=',', encoding='UTF-8')
empresa2 = pd.read_csv('empresa2.csv', sep=',', encoding='UTF-8')
empresa3 = pd.read_csv('empresa3.csv', sep=',', encoding='UTF-8')
empresa4 = pd.read_csv('empresa4.csv', sep=',', encoding='UTF-8')
empresa5 = pd.read_csv('empresa5.csv', sep=',', encoding='UTF-8')
empresa6 = pd.read_csv('empresa6.csv', sep=',', encoding='UTF-8')
empresa7 = pd.read_csv('empresa7.csv', sep=',', encoding='UTF-8')
empresa8 = pd.read_csv('empresa8.csv', sep=',', encoding='UTF-8')
empresa9 = pd.read_csv('empresa9.csv', sep=',', encoding='UTF-8')

# JUNTAR DATAFRAME EMPRESA
empresaTratada = empresa0.append(empresa1)
empresaTratada = empresaTratada.append(empresa2)
empresaTratada = empresaTratada.append(empresa3)
empresaTratada = empresaTratada.append(empresa4)
empresaTratada = empresaTratada.append(empresa5)
empresaTratada = empresaTratada.append(empresa6)
empresaTratada = empresaTratada.append(empresa7)
empresaTratada = empresaTratada.append(empresa8)
empresaTratada = empresaTratada.append(empresa9)

#EXPORTAR TABELA EMPRESA TOTAL (empresa)
empresaTratada.to_csv(r'empresaTratada.csv', index = False)

#VALIDAR REGISTROS DE EMPRESAS TOTAL
print('Empresa 0 contem ', empresa0.shape, ' registros')
print('Empresa 1 contem ', empresa1.shape, ' registros')
print('Empresa 2 contem ', empresa2.shape, ' registros')
print('Empresa 3 contem ', empresa3.shape, ' registros')
print('Empresa 4 contem ', empresa4.shape, ' registros')
print('Empresa 5 contem ', empresa5.shape, ' registros')
print('Empresa 6 contem ', empresa6.shape, ' registros')
print('Empresa 7 contem ', empresa7.shape, ' registros')
print('Empresa 8 contem ', empresa8.shape, ' registros')
print('Empresa 9 contem ', empresa9.shape, ' registros')
print('Total de registros = ', empresa0.shape[0] + empresa1.shape[0] 
+ empresa2.shape[0] + empresa3.shape[0] + empresa4.shape[0] + empresa5.shape[0] 
+ empresa6.shape[0] +empresa7.shape[0] + empresa8.shape[0] + empresa9.shape[0])
print('Empresas Total contem ', empresaTratada.shape, ' registros')

empresa0.head()


empresa1.head()

empresa2.head()

empresa3.head()

empresa4.head()

empresa5.head()

empresa6.head()

empresa7.head()

empresa8.head()

empresa9.head()

#VERIFICAR DADOS FALTANTES
print(pd.isnull(empresaTratada).sum())

#TESTANTO INTEGRIDADE DO ARQUIVO EMPRESAS
os.chdir(tratada)
empresaTratada = pd.read_csv('empresaTratada.csv', sep=',', encoding='UTF-8', usecols = ['CNPJ BASICO','NATJU','CAP_SOCIAL','QUALIF','PORTE'])

empresaTratada.shape

empresaTratada.info()

empresaTratada.head()

#LIMPAR MEMORIA
del empresaTratada
import gc
gc.collect()

TRATAMENTO DE DADOS INICIAL ARQUIVOS ESTABELECIMENTO DA BASE RFB  K3241.K03200Y0.D20312.ESTABELE.CSV A K3241.K03200Y10.D20312.ESTABELE.CSV

#IMPORTAR TABELA ESTABELECIMENTOS #0
os.chdir(caminho)
estab0 = pd.read_csv('K3241.K03200Y0.D20312.ESTABELE.CSV', sep=';', encoding='ISO-8859-1')
estab0.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 0 
os.chdir(tratada)
estab0.to_csv(r'estab0.csv', index = False)

#LIMPAR MEMORIA
del estab0
import gc
gc.collect()

#IMPORTAR TABELA ESTABELECIMENTOS #1
os.chdir(caminho)
estab1 = pd.read_csv('K3241.K03200Y1.D20312.ESTABELE.CSV', sep=';', encoding='ISO-8859-1')
estab1.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 1 
os.chdir(tratada)
estab1.to_csv(r'estab1.csv', index = False)

#LIMPAR MEMORIA
del estab1
import gc
gc.collect()

#IMPORTAR TABELA ESTABELECIMENTOS #2
os.chdir(caminho)
estab2 = pd.read_csv('K3241.K03200Y2.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab2.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 2 
os.chdir(tratada)
estab2.to_csv(r'estab2.csv', index = False)

#LIMPAR MEMORIA
del estab2
import gc
gc.collect()


#IMPORTAR TABELA ESTABELECIMENTOS #3
os.chdir(caminho)
estab3 = pd.read_csv('K3241.K03200Y3.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab3.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 3 
os.chdir(tratada)
estab3.to_csv(r'estab3.csv', index = False)

#LIMPAR MEMORIA
del estab3
import gc
gc.collect()

#IMPORTAR TABELA ESTABELECIMENTOS #4
os.chdir(caminho)
estab4 = pd.read_csv('K3241.K03200Y4.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab4.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 4 
os.chdir(tratada)
estab4.to_csv(r'estab4.csv', index = False)

#LIMPAR MEMORIA
del estab4
import gc
gc.collect()

#IMPORTAR TABELA ESTABELECIMENTOS #5
os.chdir(caminho)
estab5 = pd.read_csv('K3241.K03200Y5.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab5.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 5 
os.chdir(tratada)
estab5.to_csv(r'estab5.csv', index = False)

#LIMPAR MEMORIA
del estab5
import gc
gc.collect()


#IMPORTAR TABELA ESTABELECIMENTOS #6
os.chdir(caminho)
estab6 = pd.read_csv('K3241.K03200Y6.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab6.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 6
os.chdir(tratada)
estab6.to_csv(r'estab6.csv', index = False)

#LIMPAR MEMORIA
del estab6
import gc
gc.collect()


#IMPORTAR TABELA ESTABELECIMENTOS #7
os.chdir(caminho)
estab7 = pd.read_csv('K3241.K03200Y7.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab7.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 7
os.chdir(tratada)
estab7.to_csv(r'estab7.csv', index = False)

#LIMPAR MEMORIA
del estab7
import gc
gc.collect()

#IMPORTAR TABELA ESTABELECIMENTOS #8
os.chdir(caminho)
estab8 = pd.read_csv('K3241.K03200Y8.D20312.ESTABELE.csv', sep=';', encoding='ISO-8859-1')
estab8.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 8
os.chdir(tratada)
estab8.to_csv(r'estab8.csv', index = False)

#LIMPAR MEMORIA
del estab8
import gc
gc.collect()

#IMPORTAR TABELA ESTABELECIMENTOS #9
os.chdir(caminho)
estab9 = pd.read_csv('K3241.K03200Y9.D20312.ESTABELE.CSV', sep=';', encoding='ISO-8859-1')
estab9.columns = ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','NOME_FANT','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','NOME_DA_CID','PAIS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','TIPO_DE_LOG','LOGRADOURO','NUMERO','COMPL','BAIRRO','CEP','UF','MUNICIPIO','DDD_1','TELEFONE_1','DDD_2','TELEFONE_2','DDD_DO_FAX','FAX','CORREIO_ELET','SITUACAO_ESPECIAL','DATA_DA_SIT']

#EXPORTAR TABELA ESTAB 9
os.chdir(tratada)
estab9.to_csv(r'estab9.csv', index = False)

#LIMPAR MEMORIA
del estab9
import gc
gc.collect()


os.chdir(tratada)
#IMPORTAR ESTABELECIMENTO 
estab0 = pd.read_csv('estab0.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab1 = pd.read_csv('estab1.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab2 = pd.read_csv('estab2.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab3 = pd.read_csv('estab3.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab4 = pd.read_csv('estab4.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab5 = pd.read_csv('estab5.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab6 = pd.read_csv('estab6.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab7 = pd.read_csv('estab7.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab8 = pd.read_csv('estab8.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])
estab9 = pd.read_csv('estab9.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','CNAE_FISCAL_SEC','UF','MUNICIPIO'])



# JUNTAR DATAFRAME ESTABELECIMENTO

estabTratada = estab0.append(estab1)
estabTratada = estabTratada.append(estab2)
estabTratada = estabTratada.append(estab3)
estabTratada = estabTratada.append(estab4)
estabTratada = estabTratada.append(estab5)
estabTratada = estabTratada.append(estab6)
estabTratada = estabTratada.append(estab7)
estabTratada = estabTratada.append(estab8)
estabTratada = estabTratada.append(estab9)

# VERIFICAR INTEGRIDADE DE REGISTROS DE ESTABELECIMENTO
print('Estab 0 contem ', estab0.shape, ' registros')
print('Estab 1 contem ', estab1.shape, ' registros')
print('Estab 2 contem ', estab2.shape, ' registros')
print('Estab 3 contem ', estab3.shape, ' registros')
print('Estab 4 contem ', estab4.shape, ' registros')
print('Estab 5 contem ', estab5.shape, ' registros')
print('Estab 5 contem ', estab6.shape, ' registros')
print('Estab 5 contem ', estab7.shape, ' registros')
print('Estab 8 contem ', estab8.shape, ' registros')
print('Estab 9 contem ', estab9.shape, ' registros')
print('Total de registros = ', estab0.shape[0] + estab1.shape[0] + estab2.shape[0] + estab3.shape[0] + estab4.shape[0] + estab5.shape[0] + estab6.shape[0] + estab7.shape[0] + estab8.shape[0] + estab9.shape[0])
print('Estabs Total contem ', estabTratada.shape, ' registros')

estab0.head()

estab1.head()

estab2.head()

estab3.head()

estab4.head()

estab5.head()

estab6.head()

estab7.head()

estab8.head()

estab9.head()

# verifica se existe qualquer variável com dados faltantes
print(pd.isnull(estabTratada).sum())

estabTratada.shape

estabTratada.info()

estabTratada.head()

#EXPORTAR TABELA EMPRESA TOTAL (empresa)
estabTratada.to_csv(r'estabTratada.csv', index = False)

#LIMPAR MEMORIA
del estabTratada
del estab0
del estab1
del estab2
del estab3
del estab4
del estab5
del estab6
del estab7
del estab8
del estab9
del empresa0
del empresa1
del empresa2
del empresa3
del empresa4
del empresa5
del empresa6
del empresa7
del empresa8
del empresa9
import gc
gc.collect()
