==> 02-Agregar.ipynb <==

#Sobrevivência empresarial em Pernambuco com base no PIB e nos dados abertos da RFB: uma análise em dois períodos sociopolítico econômicos da história do Brasil

##Elisângela P.M. Santana¹, Edivaldo R.S. Júnior², Cleide Lins de Almeida³
##SCRIPT 02 - AGREGAR

#IMPORTACAO DAS BIBLIOTECAS 
import os
import numpy as np
import pandas as pd
os.getcwd()

# IMPORTACAO DA TABELA EMPRESA
#cnpj_basico, #capital_social, #porte_empresa #Natureza Juridica

os.chdir('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA')
empresaTratada = pd.read_csv('empresaTratada.csv', sep=',', encoding='UTF-8', usecols = ['CNPJ BASICO','NATJU','CAP_SOCIAL','PORTE'])
empresaTratada.head()

#RENOMEAR COLUNA CNPJ BASICO
empresaTratada = empresaTratada.rename(columns={'CNPJ BASICO': 'CNPJ_BASICO'})

empresaTratada.shape

#VERIFICAR VALORES FALTANTES
print(pd.isnull(empresaTratada).sum())

#TAMANHO DA TABELA EMPRESA
empresaTratada.shape[0]

empresaTratada.dtypes

# TRANSFORMAR COLUNA CAPITAL SOCIAL EM INTEIRO
# PASSO 01 - MUDAR VIRGULA PARA PONTO NA COLUNA CAP_SOCIAL
def mudarNumero(item):
  if ',' in item:
    return item.replace(',','.')
  else:
    return item

empresaTratada['CAP_SOCIAL'] = empresaTratada['CAP_SOCIAL'].apply(mudarNumero)
empresaTratada['CAP_SOCIAL'].head()

#PASSO 02 TRANSFORMAR CAP_SOCIAL EM NUMERICO (float64)
empresaTratada['CAP_SOCIAL'] = pd.to_numeric(empresaTratada['CAP_SOCIAL'])
empresaTratada.dtypes

print('Quantidade de registros',empresaTratada.shape[0])
print('Tipo de Registros','\n',empresaTratada.dtypes)

#EXPORTAR empresaTratada FORMATADA
empresaTratada.to_csv(r'empresaFormatada.csv', index = False)

del empresaTratada

# IMPORTACAO DA TABELA ESTABELECIMENTO
#cnpj_basico, ordem , dv  #Matriz ou Filial #sit_cadastral, #dt_sit_cadastral, #Mot_situação cadastral #dt_ini_atividade, #cnae_fisc_princ, #UF #municipio
os.chdir('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA')
estabTratada = pd.read_csv('estabTratada.csv', sep=',', encoding='UTF-8', usecols= ['CNPJ_BASICO','CNPJ_ORDEM','CNPJ_DV','IDE_MAT_FIL','SIT_CADASTRAL','DATA_SIT_CADAS','MOT_SIT_CADAS','DATA_INI_ATV','CNAE_FISCAL','UF','MUNICIPIO'])

#TAMANHO DA TABELA ESTABELECIMENTO
estabTratada.shape[0]

#QUANTIDADE DE MATRIZ E FILIAL
estabTratada['IDE_MAT_FIL'].value_counts()

#VERIFICAR VALORES FALTANTES
print(pd.isnull(estabTratada).sum())

# VERIFICANDO ESTRUTURA DA TABELA ESTABELECIMENTO
estabTratada.head()

#VERIFICANDO TYPO DA TABELA ESTABELECIMENTO
estabTratada.dtypes

#IMPORTAR TABELA CNAE
cnae =  pd.read_csv('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA\cnae.csv', sep=',', encoding='UTF-8')

#TIPO DA TABELA CNAE
cnae.dtypes

# IMPORTAR empresaTratada FORMATADA
empresaFormatada =  pd.read_csv('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA\empresaFormatada.csv', sep=',', encoding='UTF-8')


empresaFormatada.info()

estabTratada.info()


#JUNTAR DATASET EMPRESA E ESTABELECIMENTO
empresaFormatada.info()
estabTratada.info()


#JUNTAR DATASET EMPRESA E ESTABELECIMENTO
empresasBrasil = pd.merge(empresaFormatada, estabTratada, how = 'outer', on = 'CNPJ_BASICO')

#VERIFICAR VALORES FALTANTES
print(pd.isnull(empresasBrasil).sum())

#EXPORTAR TABELA EMPRESAS BRASIL (Empresa + Estabelecimento)
os.chdir('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA')
empresasBrasil.to_csv(r'empresasBrasil.csv', index = False)


print('Registros Base Estabelecimentos', estabTratada.shape)
print('Registros Base Empresas Brasil', empresasBrasil.shape)
print('Diferença de registros', empresasBrasil.shape[0] - estabTratada.shape[0])

#QUANTIDADE DE MATRIZ E FILIAL
print('Quantidade de Matriz e Filiais','\n',empresasBrasil['IDE_MAT_FIL'].value_counts(),'\n')


#VERIFICAR VALORES FALTANTES
print(pd.isnull(empresasBrasil).sum())

#IMPORTAR TABELA SIMPLES PARA IDENTIFICAÇÃO DO MEI
os.chdir('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA')
simples = pd.read_csv('simples.csv', sep=',', encoding='UTF-8', usecols=['CNPJ BASICO', 'OPÇÃO PELO SIMPLES','MEI'])



simples.tail()

simples = simples.rename(columns={'CNPJ BASICO': 'CNPJ_BASICO','OPÇÃO PELO SIMPLES': 'OP_SIMPLES'})

simples.info()

#JUNTAR DATASERT EMPRESA (Empresa + Estabelecimento + Simples)
empresasBrasil = pd.merge(empresasBrasil, simples, how = 'outer', on = 'CNPJ_BASICO')

#EXPORTAR TABELA EMPRESAS BRASIL (Empresa + Estabelecimento + Simples)
os.chdir('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA')
empresasBrasil.to_csv(r'empresasBrasil02.csv', index = False)

empresasBrasil.shape

empresasBrasil.head()

empresasBrasil.info()

empresasBrasil.describe()

#VERIFICAR VALORES FALTANTES
print(pd.isnull(empresasBrasil).sum())

empresasBrasil.info()
