==> 03-TratamentoFinal.ipynb <==
## Sobrevivência empresarial em Pernambuco com base no PIB e nos dados abertos da RFB: uma análise em dois períodos sociopolítico econômicos da história do Brasil


#Elisângela P.M. Santana¹, Edivaldo R.S. Júnior², Cleide Lins de Almeida³
## SCRIPT 03 - TRATAMENTO FINAL

# 1.0 Carregando Bibliotecas

# bibliotecas para análise de dados e para visualização
import numpy as np
import pandas as pd
import xlrd
import os
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
# configurando variáveis do ambiente
import warnings
warnings.filterwarnings('ignore')
pd.options.display.max_columns = None
pd.options.display.max_rows = None
pd.options.display.float_format = '{:.4f}'.format

# 2.0 Visualização e tratamento dataset empresas Brasil 

from google.colab import drive
drive.mount('/content/drive')

# IMPORTACAO DA TABELA EMPRESAS BRASIL
caminho = ('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA')
os.chdir(caminho)           
empresasBrasil = pd.read_csv('empresasBrasil.csv', sep=',', encoding='UTF-8')

# INFORMAÇÕES DO DATASET EMPRESAS BRASIL
empresasBrasil.info()

# VISUALIZAÇÃO DO DATASET EMPRESAS BRASIL
empresasBrasil.head(2)

#TAMANHO DO DATASET
empresasBrasil.shape

#SEPARAR EMPRESAS DE PERNAMBUCO 
def listaPib(dataframe, valores):
    return dataframe.loc[dataframe['UF'].isin(valores)]
lista = ['PE']
empresasPE = listaPib(empresasBrasil, lista)
empresasPE.shape

empresasPE.info()

#EXPORTAR TABELA DE EMPRESAS DE PE
empresasPE.to_csv(r'D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA\empresasPE.csv', index = False)

#IMPORTAR TABELA DE EMPRESAS DE PE
empresasPE = pd.read_csv('D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA\empresasPE.csv', sep=',', encoding='UTF-8')

#VERIFICAR VALORES FALTANTES
print(pd.isnull(empresasPE).sum())

# VERIFICAR DADOS DUPLICADOS
duplicados = empresasPE[empresasPE.duplicated(keep='first')]
print(duplicados)

duplicados.head()

#TRATAMENTO 

# FORMATAR NUMERO DO MUNICIPIO PARA INTEIRO 
empresasPE['MUNICIPIO'] = empresasPE['MUNICIPIO'].astype('int')
empresasPE['MUNICIPIO'].dtypes

#VISUALIZAR DATASET
empresasPE.head()

# EXCLUIR MEI
empresasPEsemMei = empresasPE[empresasPE['MEI']!='S']

print("sem MEI = ", empresasPEsemMei.shape)
print("com MEI = ", empresasPE.shape)
print("Total de MEI =", empresasPE.shape[0] - empresasPEsemMei.shape[0])

#VERIFICAR CNAES FISCAIS - AGREGAR AO DATASET EMPRESAS
empresasPEsemMei['CNAE_FISCAL'].value_counts().head()

cnae = pd.read_csv('cnae.csv', sep=',',encoding='UTF-8')

cnae.head()

cnae = cnae.rename(columns={'N_CNAE': 'CNAE_FISCAL'})

empresasPEsemMei = pd.merge(empresasPEsemMei,cnae, on='CNAE_FISCAL')

empresasPEsemMei.info()

#ANALISAR CNAES DAS EMPRESAS DE PE
empresasPEsemMei["DESC_CNAE"].value_counts().head()

# EXCLUIR CNAE ['Atividades de organizações políticas', 'Atividade Econônica não informada']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_CNAE"]!= "Atividades de organizações políticas"]
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_CNAE"]!="Atividade Econônica não informada"]
empresasPEsemMei["DESC_CNAE"].value_counts().head()

#ANALISAR AS NATUREZAS JURIDICAS 
empresasPEsemMei["NATJU"].value_counts().head()

#IMPORTAR E AGREGAR NATUREZA JURIDICA AO DATASET
natju = pd.read_csv('natju.csv', sep=",", encoding="UTF-8")

natju.head()

natju = natju.rename(columns={'N_NATJU': 'NATJU'})

empresasPEsemMei = pd.merge(empresasPEsemMei,natju, on='NATJU')

# naturezas juridicas para Excluir

empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Organização Religiosa']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Executivo Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundo Público da Administração Direta Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Empresa Pública']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Executivo Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundação Pública de Direito Público Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Município']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Autarquia Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Executivo Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Legislativo Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Autarquia Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Autarquia Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundação Pública de Direito Público Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundação Pública de Direito Público Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Organização Social (OS)']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Consórcio Público de Direito Público (Associação Pública)']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público Autônomo Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundo Público da Administração Direta Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão de Direção Local de Partido Político']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Estabelecimento, no Brasil, de Fundação ou Associação Estrangeiras']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Representação Diplomática Estrangeira']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Judiciário Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Estado ou Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Judiciário Estadual']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público Autônomo Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundo Público da Administração Direta Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Legislativo Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Consórcio Público de Direito Privado']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundo Público da Administração Indireta Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão de Direção Regional de Partido Político']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público Autônomo Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Órgão Público do Poder Legislativo Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Comunidade Indígena']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundação Pública de Direito Privado Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundo Público da Administração Indireta Estadual ou do Distrito Federal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Fundação Pública de Direito Privado Municipal']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Outras Instituições Extraterritoriais']
empresasPEsemMei = empresasPEsemMei[empresasPEsemMei["DESC_NATJU"]!='Entidade Sindical']
empresasPEsemMei["DESC_NATJU"].value_counts().head()

empresasPEsemMei.shape

#AGREGAR MUNICIPIOS A TABELA DE EMPRESAS 
municipiorfb = pd.read_csv('RFBMunicipio.csv', sep=";", encoding="UTF-8")

municipiorfb.head()

municipiorfb.columns

municipiorfb = municipiorfb.rename(columns={'Código Município IBGE': 'MUNICIPIO'})

municipiorfbPe = municipiorfb[municipiorfb['Sigla UF'] == "PE"]

municipiorfbPe.head()

municipiorfbPe['MUNICIPIO'] = municipiorfbPe['MUNICIPIO'].astype('int')

municipiorfbPe.head()

caminho = 'D:\\#PROJETOS\\FURG\\ECD-CIENCIA DE DADOS\\#TCC-FURG\\BASE-RFB-12-03-2022'
tratada = 'D:\#PROJETOS\FURG\ECD-CIENCIA DE DADOS\#TCC-FURG\BASE-RFB-12-03-2022\TRATADA'
#EXPORTAR TABELA MUNICIPIO RFB
os.chdir(tratada)
municipiorfbPe.to_csv(r'municipiorfbPe.csv', index = False)

#IMPORTAR TABELA MUNICIPIO RFB SELECIONADA
os.chdir(tratada)
municipiorfbPe = pd.read_csv('municipiorfbPe.csv', sep=',', encoding='UTF-8', usecols=['Nome Município','Latitude Município','Longitude Município','Código Município SRF','MUNICIPIO'])
municipiorfbPe.head()

municipiorfbPe['Código Município SRF'] = municipiorfbPe['Código Município SRF'].astype('int')

municipiorfbPe.head()

empresasPEsemMei.head()

municipiorfbPe.head()

#municipiorfbPe = municipiorfbPe.rename(columns={'Código Município SRF': 'N_MUNIC'})
empresasPEsemMei = empresasPEsemMei.rename(columns={'MUNICIPIO': 'N_MUNIC'})

#MESCLAR TABELA MUNICIPIO IBGE E EMPRESAS
empresasPEsemMei = pd.merge(empresasPEsemMei,municipiorfbPe, on='N_MUNIC')

empresasPEsemMei.head()

empresasPEsemMei = empresasPEsemMei.rename(columns={'Nome Município': 'NOME'})
empresasPEsemMei = empresasPEsemMei.rename(columns={'Latitude Município': 'LATITUDE'})
empresasPEsemMei = empresasPEsemMei.rename(columns={'Longitude Município': 'LONGITUDE'})

empresasPEsemMei.columns

empresasPEsemMei['PORTE'].value_counts()

empresasPEsemMei.shape

df_FinalPe = empresasPEsemMei

df_FinalPe.isnull().sum()

# FORMATAR NUMERO  INTEIRO 
df_FinalPe['NATJU'] = df_FinalPe['NATJU'].astype('int')
df_FinalPe['CAP_SOCIAL'] = df_FinalPe['CAP_SOCIAL'].astype('int')
df_FinalPe['IDE_MAT_FIL'] = df_FinalPe['IDE_MAT_FIL'].astype('int')
df_FinalPe['SIT_CADASTRAL'] = df_FinalPe['SIT_CADASTRAL'].astype('int')
df_FinalPe['MOT_SIT_CADAS'] = df_FinalPe['MOT_SIT_CADAS'].astype('int')
df_FinalPe['CNAE_FISCAL'] = df_FinalPe['CNAE_FISCAL'].astype('int')
df_FinalPe['DATA_INI_ATV'] = df_FinalPe['DATA_INI_ATV'].astype('int')
df_FinalPe['DATA_SIT_CADAS'] = df_FinalPe['DATA_SIT_CADAS'].astype('int')
df_FinalPe['DATA_INI_ATV'] = df_FinalPe['DATA_INI_ATV'].astype('str')
df_FinalPe['DATA_SIT_CADAS'] = df_FinalPe['DATA_SIT_CADAS'].astype('str')
df_FinalPe['PORTE'] = df_FinalPe['PORTE'].astype('int')

df_FinalPe.head()

#SEPARAR ANO DE INICIO DE ATIVIDADE E SITUAÇÃO CADASTRAL

df_FinalPe['ANO_SIT_CADAS'] = df_FinalPe['DATA_SIT_CADAS'].str.slice(0, 4)
df_FinalPe['ANO_INI_ATV'] = df_FinalPe['DATA_INI_ATV'].str.slice(0, 4)
df_FinalPe['ANO_SIT_CADAS'] = df_FinalPe['ANO_SIT_CADAS'].astype('int')
df_FinalPe['ANO_INI_ATV'] = df_FinalPe['ANO_INI_ATV'].astype('int')

df_FinalPe['PORTE'].value_counts()

df_FinalPe['ANO_INI_ATV'].value_counts(ascending=True)

df_FinalPe['PORTE'] = df_FinalPe['PORTE'].astype('str')

df_FinalPe['ANO_SIT_CADAS'].value_counts(ascending=True)

#EXCLUIR ANOS ZERADOS (2110 registros)
df_FinalPe =df_FinalPe[df_FinalPe['ANO_SIT_CADAS'] != 0]

df_FinalPe['ANO_SIT_CADAS'].value_counts(ascending=True)

df_FinalPe.columns

selec  = ['NOME','MUNICIPIO','UF','DESC_NATJU','NATJU','PORTE','IDE_MAT_FIL','SIT_CADASTRAL','MOT_SIT_CADAS','ANO_SIT_CADAS','ANO_INI_ATV',
       'CNAE_FISCAL','N_MUNIC','DESC_CNAE','LATITUDE', 'LONGITUDE']
df_FinalPe = df_FinalPe.filter(items=selec)
df_FinalPe.info()

df_FinalPe.head()

#MUDAR SITUAÇÃO CADASTRAL PARA ATIVA E INATIVA

def define_classe(num):
    if num != 2:
        return 'INATIVA'
    else:
        return 'ATIVA'
   

df_FinalPe['SIT_CADASTRAL'] = df_FinalPe['SIT_CADASTRAL'].map(define_classe)

df_FinalPe['SIT_CADASTRAL'].value_counts()

df_FinalPe.to_csv(r'df_PeFinal.csv', index = False)

df_FinalPe=pd.read_csv('/content/drive/MyDrive/FURG-TCC/TCC - DADOS E ANALISE/DADOS/FINAL/df_PeFinal.csv', sep=',', encoding='UTF-8')

df_FinalPe.info()
